c-list: use uintptr_t for pointer arithmetic in c_list_entry()
This seems more portable. Otherwise, Visual Studio 2017 warns:

  C:\...\c-list.h(91): warning C4311: 'type cast': pointer truncation from 'void *' to 'unsigned long'
  C:\...\c-list.h(91): warning C4312: 'type cast': conversion from 'unsigned long' to 'void *' of greater size

+++ src/c-list.h
@@ -22,6 +22,7 @@ extern "C" {
 #endif
 
 #include <stddef.h>
+#include <stdint.h>
 
 typedef struct CList CList;
 
@@ -80,7 +81,7 @@ static inline void *c_list_entry_offset(const CList *what, unsigned long offset)
              * dereferenced, this is fine. We explicitly use integer arithmetic
              * to circumvent the Gcc warning.
              */
-            return (void *)(((unsigned long)(void *)what) - offset);
+            return (void *)(((uintptr_t)(void *)what) - offset);
         }
         return NULL;
 }
